requirements.txt
schedule==1.1.0
pyTelegramBotAPI==4.10.0
Pillow==9.4.0